changed base type for neural network from double to float

+++ Library/NeuralNetwork/ActivationFunctions/IActivationFunction.cs
@@ -1,6 +1,6 @@
 namespace GradientDescentSharp.NeuralNetwork;
 
 public interface IActivationFunction{
-    double Activation(double x);
-    double ActivationDerivative(double x);
+    float Activation(float x);
+    float ActivationDerivative(float x);
 }
